requirements.txt
# Core dependencies for Aurora AI
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0

# Optional dependencies for extended functionality
# Uncomment as needed:
# PyGithub>=1.59.0
# twython>=3.9.1
# fastapi>=0.104.0
# uvicorn>=0.24.0
# passlib>=1.7.4
# cryptography>=41.0.0
# transformers>=4.35.0
# sentence-transformers>=2.2.2
# scikit-learn>=1.3.0
# apscheduler>=3.10.0
# bleach>=6.1.0
# cachetools>=5.3.0
# slowapi>=0.1.9

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0